Add files via upload
These are Version two of the project originally created back in 2021 and then abandoned 1/20/22, version 3 proceeds with this version and includes a method to secure the custom-cut 60T/40R glass that we ordered to complete this project. for further information, the 60T/40R glass is just a teleprompter glass 60% transparent and 40% reflective. We custom-ordered the glass from https://telepromptermirror.com/ and ordered a 6x6 sample and then took it to a local shop that cut it to specification.